requirements.txt
RPi.GPIO
flask
asyncio